[feat] navigating using local storage with wallet address
