+++ frontend/gostarkme-web/package.json
@@ -18,7 +18,8 @@
     "react": "^18",
     "react-dom": "^18",
     "starknet": "^6.7.0",
-    "starknetkit": "^1.1.5"
+    "starknetkit": "^1.1.5",
+    "usehooks-ts": "^3.1.0"
   },
   "devDependencies": {
     "@types/node": "^20",
